llm-service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
google-generativeai==0.5.4
pydantic==2.7.2